# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dispatch controller API
  version: 0.1.0
  description: Project-scoped application deployment and operations API.
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /private-networks:
    get:
      operationId: listEdgeNodes
      responses:
        "200":
          description: Edge nodes, managed Laneway Connectors, and legacy Laneway routes.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/EdgeNode" } }
    post:
      operationId: createPrivateNetwork
      description: Creates an edge node or Laneway connection. Edge credentials are returned once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
                driver: { type: string, enum: [dispatch_agent, laneway_connector, laneway], default: dispatch_agent }
                socketPath: { type: string, description: Used only by a legacy Laneway route. }
                authority: { type: string, format: uri, description: Laneway control-plane HTTPS origin used by a managed Connector. }
                route: { type: string, description: Private Dispatch IP or CIDR published through the managed Connector. }
      responses:
        "201": { description: Private network connection created. }
  /private-networks/{edgeNodeId}/install-connector:
    parameters:
      - { in: path, name: edgeNodeId, required: true, schema: { type: string } }
    post:
      operationId: installLanewayConnector
      description: Consumes a one-time Laneway Connector bootstrap and deploys the Connector through the controller Docker socket. The command is not stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bootstrapCommand]
              properties:
                bootstrapCommand: { type: string, writeOnly: true, maxLength: 4096 }
      responses:
        "200": { description: Connector started. }
        "400": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    EdgeNode:
      type: object
      required: [id, name, driver, config, details, state, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        driver: { type: string, enum: [dispatch_agent, laneway_connector, laneway, laneway_network] }
        config: { type: object, additionalProperties: { type: string } }
        details: { type: object, additionalProperties: { type: string } }
        enrollmentToken: { type: string, writeOnly: true, description: Returned only on create or rotation. }
        credentialsConfigured: { type: boolean, description: True for a connected Laneway network. }
        state: { type: string, enum: [waiting, ready, offline, error, unverified] }
        lastVerifiedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [status, title, detail]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }